+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 ## Release Notes
 
+
+---
+
+`1.0.3`
+
+- `Reflection.findSubclass` no longer throws `IllegalStateException` it throws `ReflectionException` instead.
+- Added `Nullables.requireNull` for easier preconditions throwing an `IllegalArgumentException` if the parameter is not `null`.
+- Added `JavaObjects.max` that returns the maximum of two `Comparable`s
+- Added `JavaObjects.min` that returns the minimum of two `Comparable`s
+- Updated parent to `3.4.5`
+- Added `Methods.getCallerMethodName` with skipFrames parameter.
+- Added `Consumers` class with utility methods for consumers.
+- Added `Consumers.noBiConsumer` method that returns an empty `BiConsumer`.
+- Added `Runnables` class with utility methods for runnables.
+- Removed consumer and runnable methods from `Threads` class.
+
 ---
 
 `1.0.2`